I3D & Transformer Testing
Reduce the dataset to the popular 100 in the gloss. The original 2000 has sparse data on the lower end where there are 2-3 examples for each instead of 10+ in the top 100.

for transformer created new dataset for just the 100 and new method to get keypoints. this also included testing of pose, pose hands, and pose hands face.

new dataset for I3D as well.